// apps/api/src/generated/prisma/models/Post.ts
export type PostUncheckedCreateInput = {
  id?: string
  campaignId: string
  ctaId?: string | null
  platform: $Enums.Platform
  postType: $Enums.PostType
  content: string
  hashtags?: Prisma.PostCreatehashtagsInput | string[]
  publishDate?: Date | string | null
  scheduledAt?: Date | string | null
  publishedAt?: Date | string | null
  status?: $Enums.PostStatus
  errorLog?: string | null
  createdAt?: Date | string
  updatedAt?: Date | string
  postMedia?: Prisma.PostMediaUncheckedCreateNestedManyWithoutPostInput
  analytics?: Prisma.PostAnalyticsUncheckedCreateNestedManyWithoutPostInput
}
// ...
export type PostUncheckedUpdateInput = {
  id?: Prisma.StringFieldUpdateOperationsInput | string
  campaignId?: Prisma.StringFieldUpdateOperationsInput | string
  ctaId?: Prisma.NullableStringFieldUpdateOperationsInput | string | null
  platform?: Prisma.EnumPlatformFieldUpdateOperationsInput | $Enums.Platform
  postType?: Prisma.EnumPostTypeFieldUpdateOperationsInput | $Enums.PostType
  content?: Prisma.StringFieldUpdateOperationsInput | string
  hashtags?: Prisma.PostUpdatehashtagsInput | string[]
  publishDate?: Prisma.NullableDateTimeFieldUpdateOperationsInput | Date | string | null
  scheduledAt?: Prisma.NullableDateTimeFieldUpdateOperationsInput | Date | string | null
  publishedAt?: Prisma.NullableDateTimeFieldUpdateOperationsInput | Date | string | null
  status?: Prisma.EnumPostStatusFieldUpdateOperationsInput | $Enums.PostStatus
  errorLog?: Prisma.NullableStringFieldUpdateOperationsInput | string | null
  createdAt?: Prisma.DateTimeFieldUpdateOperationsInput | Date | string
  updatedAt?: Prisma.DateTimeFieldUpdateOperationsInput | Date | string
  postMedia?: Prisma.PostMediaUncheckedUpdateManyWithoutPostNestedInput
  analytics?: Prisma.PostAnalyticsUncheckedUpdateManyWithoutPostNestedInput
}
// ...
export type PostCreatehashtagsInput = {
  set: string[]
}

export type EnumPlatformFieldUpdateOperationsInput = {
  set?: $Enums.Platform
}

export type EnumPostTypeFieldUpdateOperationsInput = {
  set?: $Enums.PostType
}

export type PostUpdatehashtagsInput = {
  set?: string[]
  push?: string | string[]
}

export type EnumPostStatusFieldUpdateOperationsInput = {
  set?: $Enums.PostStatus
}